Base de Datos MySQL Workbench/requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
mysql-connector-python==9.0.0
pandas==2.2.2
python-dotenv==1.0.1
